requirements.txt
gspread == 5.6.0
google-api-python-client == 2.64.0
qrcode==7.3.1
pillow==9.2.0